+++ crates/core/config/src/project/workspace.rs
@@ -1,6 +1,7 @@
 // These configs are project-level settings that override those from the workspace!
 
 use crate::types::TaskID;
+use crate::validators::is_default;
 use rustc_hash::FxHashMap;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -10,10 +11,13 @@ use validator::Validate;
 #[schemars(default)]
 #[serde(default)]
 pub struct ProjectWorkspaceInheritedTasksConfig {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub exclude: Option<Vec<TaskID>>,
 
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub include: Option<Vec<TaskID>>,
 
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub rename: Option<FxHashMap<TaskID, TaskID>>,
 }
 
@@ -21,6 +25,7 @@ pub struct ProjectWorkspaceInheritedTasksConfig {
 #[schemars(default)]
 #[serde(default, rename_all = "camelCase")]
 pub struct ProjectWorkspaceConfig {
+    #[serde(skip_serializing_if = "is_default")]
     #[validate]
     pub inherited_tasks: ProjectWorkspaceInheritedTasksConfig,
 }