+++ crates/core/config/src/workspace/generator.rs
@@ -1,3 +1,4 @@
+use crate::validators::is_default;
 use crate::{
     errors::create_validation_error, types::FilePath, validators::validate_child_relative_path,
 };
@@ -24,8 +25,9 @@ fn validate_templates(files: &[FilePath]) -> Result<(), ValidationError> {
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Validate)]
 #[schemars(default)]
-#[serde(rename_all = "camelCase")]
+#[serde(default, rename_all = "camelCase")]
 pub struct GeneratorConfig {
+    #[serde(skip_serializing_if = "is_default")]
     #[validate(custom = "validate_templates")]
     pub templates: Vec<FilePath>,
 }