+++ crates/core/config/src/workspace/notifier.rs
@@ -12,8 +12,9 @@ fn validate_webhook_url(url: &str) -> Result<(), ValidationError> {
 
 #[derive(Clone, Debug, Default, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Validate)]
 #[schemars(default)]
-#[serde(rename_all = "camelCase")]
+#[serde(default, rename_all = "camelCase")]
 pub struct NotifierConfig {
+    #[serde(skip_serializing_if = "Option::is_none")]
     #[validate(custom = "validate_webhook_url")]
     pub webhook_url: Option<String>,
 }